--- a/mw_context_vars.go
+++ b/mw_context_vars.go
@@ -23,37 +23,24 @@ func (m *MiddlewareContextVars) EnabledForSpec() bool {
 
 // ProcessRequest will run any checks on the request on the way through the system, return an error to have the chain fail
 func (m *MiddlewareContextVars) ProcessRequest(w http.ResponseWriter, r *http.Request, _ interface{}) (error, int) {
+	r.ParseForm()
 
-	copiedRequest := copyRequest(r)
-	contextDataObject := make(map[string]interface{})
+	contextDataObject := map[string]interface{}{
+		"request_data": r.Form, // Form params (map[string][]string)
+		"headers":      map[string][]string(r.Header),
+		"headers_Host": r.Host,
+		"path_parts":   strings.Split(r.URL.Path, "/"), // Path parts
+		"path":         r.URL.Path,                     // path data
+		"remote_addr":  request.RealIP(r),              // IP
+		"request_id":   uuid.NewV4().String(),          //Correlation ID
+	}
 
-	copiedRequest.ParseForm()
-
-	// Form params (map[string][]string)
-	contextDataObject["request_data"] = copiedRequest.Form
-
-	contextDataObject["headers"] = map[string][]string(copiedRequest.Header)
-
-	for hname, vals := range copiedRequest.Header {
+	for hname, vals := range r.Header {
 		n := "headers_" + strings.Replace(hname, "-", "_", -1)
 		contextDataObject[n] = vals[0]
 	}
-	contextDataObject["headers_Host"] = copiedRequest.Host
 
-	// Path parts
-	segmentedPathArray := strings.Split(copiedRequest.URL.Path, "/")
-	contextDataObject["path_parts"] = segmentedPathArray
-
-	// path data
-	contextDataObject["path"] = copiedRequest.URL.Path
-
-	// IP
-	contextDataObject["remote_addr"] = request.RealIP(copiedRequest)
-
-	//Correlation ID
-	contextDataObject["request_id"] = uuid.NewV4().String()
-
-	for _, c := range copiedRequest.Cookies() {
+	for _, c := range r.Cookies() {
 		name := "cookies_" + strings.Replace(c.Name, "-", "_", -1)
 		contextDataObject[name] = c.Value
 	}