added changes for https://github.com/TykTechnologies/tyk/issues/1641 and https://github.com/TykTechnologies/tyk/issues/1642

changes:
- added test with three cases and benchmark for them on middleware level as it was too hard to optimize using existing benchmark with tykTestServer
- added some optimizations when we populate map before adding to context (less extra assignments and access map by key - almost all work when we initialize map)
- removed `copyRequest` call as I found that it has zero-effect in that MW but adds lots of extra work

Initially master didn't have benchmarks I've added with that PR so I did `benchcmp` of version `func (m *MiddlewareContextVars) ProcessRequest` without changes vs. version with changes:

```
benchmark                                                                                                   old ns/op     new ns/op     delta
BenchmarkContextVarsMiddlewareProcessRequest/GET_with_query_string-8                                        8866          8051          -9.19%
BenchmarkContextVarsMiddlewareProcessRequest/POST_with_query_string_and_encoded_form_data-8                 15727         12460         -20.77%
BenchmarkContextVarsMiddlewareProcessRequest/POST_with_query_string_and_encoded_form_data_and_cookies-8     16990         14973         -11.87%

benchmark                                                                                                   old allocs     new allocs     delta
BenchmarkContextVarsMiddlewareProcessRequest/GET_with_query_string-8                                        45             40             -11.11%
BenchmarkContextVarsMiddlewareProcessRequest/POST_with_query_string_and_encoded_form_data-8                 71             64             -9.86%
BenchmarkContextVarsMiddlewareProcessRequest/POST_with_query_string_and_encoded_form_data_and_cookies-8     87             80             -8.05%

benchmark                                                                                                   old bytes     new bytes     delta
BenchmarkContextVarsMiddlewareProcessRequest/GET_with_query_string-8                                        7626          7114          -6.71%
BenchmarkContextVarsMiddlewareProcessRequest/POST_with_query_string_and_encoded_form_data-8                 13066         10504         -19.61%
BenchmarkContextVarsMiddlewareProcessRequest/POST_with_query_string_and_encoded_form_data_and_cookies-8     14751         12190         -17.36%
```